requirements.txt
Flask==3.1.3
flask-cors==6.0.2
sentence-transformers==5.4.1
scikit-learn==1.7.2
numpy==2.3.3
google-genai==1.73.1
torch==2.1.0